move sidebarnotify.tsx location and record frompage to navigate
